feature: add payment refund

+++ src/app/pages/register-form/components/register-form/register-form.component.ts
@@ -135,22 +135,28 @@ export class RegisterFormComponent implements OnInit{
   }
 
   removeAttachTicket(data: any):void {
-    const formRegisterId = {
-      formRegisterId: data.id
-    }
-
-    this.registerFormService.removeAttachTicket(formRegisterId).subscribe({
-      next: res => {
-        if (res.success) {
-          this.message.success(res.messages);
-          this.pageSize$.next(10)
-        } else {
-          this.message.error(res.errorMessages)
+    this.modal.confirm({
+      nzTitle: 'Xác nhận gỡ vé',
+      nzContent: 'Bạn có chắc chắn muốn gỡ vé ' + data.ticketName + '?',
+      nzOnOk: () => {
+        const formRegisterId = {
+          formRegisterId: data.id
         }
-      },
-      error: err => {
-        this.message.error(err)
+
+        this.registerFormService.removeAttachTicket(formRegisterId).subscribe({
+          next: res => {
+            if (res.success) {
+              this.message.success(res.messages);
+              this.pageSize$.next(10)
+            } else {
+              this.message.error(res.errorMessages)
+            }
+          },
+          error: err => {
+            this.message.error(err)
+          }
+        })
       }
-    })
+    });
   }
 }
